@@ -7,7 +7,7 @@ namespace Authorization.Controllers;
 
 [ApiController]
 [Route("api/[controller]")]
-public class BaseController : ControllerBase
+public class BaseController : Controller
 {
     protected Guid UserId => Guid.Parse(User.Claims.First(c => c.Type == ClaimTypes.NameIdentifier).Value);
 }
